+++ packages/megadoc-plugin-js/CHANGES.md
@@ -1,3 +1,8 @@
+## 6.0.0-beta.9
+
+- Added a new option `inferNamespaces`  for controlling the behaviour of
+  inferring namespaces from @module tags when `.` is used
+
 ## 4.0.2
 
 - turned off the debugging messages during parse. To turn them on, run megadoc with the environment variable `NODE_DEBUG=megadoc` exported.